refactor: Service를 이용하여 API 연동 의존 구조 개선
