+++ src/main/java/roomescape/service/command/PaymentCommandService.java
@@ -1,7 +1,6 @@
 package roomescape.service.command;
 
 import org.springframework.stereotype.Service;
-import roomescape.client.PaymentClient;
 import roomescape.domain.payment.Payment;
 import roomescape.dto.payment.PaymentResponseDto;
 import roomescape.dto.reservation.PaymentConfirmDto;
@@ -9,14 +8,14 @@ import roomescape.dto.reservation.PaymentConfirmDto;
 @Service
 public class PaymentCommandService {
 
-    private final PaymentClient paymentClient;
+    private final PaymentService paymentService;
 
-    public PaymentCommandService(PaymentClient paymentClient) {
-        this.paymentClient = paymentClient;
+    public PaymentCommandService(final PaymentService paymentService) {
+        this.paymentService = paymentService;
     }
 
     public PaymentResponseDto confirmPayment(PaymentConfirmDto requestDto) {
-        Payment payment = paymentClient.confirmPayment(requestDto);
+        Payment payment = paymentService.confirmPayment(requestDto);
         return new PaymentResponseDto(payment.getOrderId(), payment.getTotalAmount());
     }
 }
