+++ src/main/java/roomescape/service/command/PaymentService.java
@@ -1,9 +1,9 @@
-package roomescape.client;
+package roomescape.service.command;
 
 import roomescape.domain.payment.Payment;
 import roomescape.dto.reservation.PaymentConfirmDto;
 
-public interface PaymentClient {
+public interface PaymentService {
 
     Payment confirmPayment(PaymentConfirmDto requestDto);
 }
